Add link from /downloads to Getting Started guide
The Getting Started guide is designed to help new users. Most traffic on the
site routes to the /downloads page today (it is 22% of overall pageviews).
Adding a prominent link from this page to the Getting Started guide increases
the visibility / discoverability of this resource in a natural and intuitive
way, and should thus drive more traffic to it, such that we get more informed
and capable new users right from their very first interactions with Bisq.