Change page titles: bisq is network not exchange
Page titles across the site call Bisq "The
decentralized Bitcoin exchange" which isn't
really true...it's more of a network that
serves as a means to exchange bitcoin.

Also "the" sounds pretentious, as if it's the
only entity of its kind that will ever exist,
so it's changed to "a".