Fix weird bug with svg images on /index
It seems the SVGs used for screenshots on the front
page did not behave consistently across OSes. Testing
indicated no issues on multiple browsers on Ubuntu,
but testing on macOS indicated SVG font properties
were not being applied correctly.

Images were PNGs before, and there was no particular need
to switch to SVG, so this commit switches them back to PNG.

The switch was initially made in #278 to produce
translated screenshots. SVG files remain in the repository
as "source" for the PNG files.