Remove press.md
It's outdated and unused. There are some links to this page
from old blog posts.

Since the press kit is largely a repository of assets like
logos and PDFs that don't change often, it makes more sense
to host these on GitHub and link to them from the website.

As soon as these assets are created and uploaded, a link to
them will be placed in the site footer, a redirect for /press
will be made to the GitHub repo, and the /resources folder will
also be removed.