Remove roadmap.md
The roadmap page has become more backward-looking than
forward-looking...it's basically a copy of release notes.

The closest thing the project has to a roadmap is the
master projects board on GitHub, so this commit redirects
/roadmap to that board.

This commit also removes extraneous links to the roadmap
in front page copy across languages.